[backend] Remove hardcoded opensaml dependencies
